+++ _clear_store.d.ts
@@ -1,7 +1,6 @@
-import type { writable_stores_val_type } from './writable_stores_val_type';
-import type { writable_stores_type } from './writable_stores_type';
+import type { Writable } from './lib';
 /**
  * Returns a function that [clear_store](#clear_store).
  */
-export declare function _clear_store<I extends writable_stores_type<unknown>>(stores: I, value: writable_stores_val_type<I>): () => void;
+export declare function _clear_store<I extends unknown>(stores: Writable<I> | Writable<I>[], value: I): () => void;
 export declare const _clear__store: typeof _clear_store;