+++ clear_store.d.ts
@@ -1,7 +1,6 @@
-import type { writable_stores_type } from './writable_stores_type';
-import type { writable_stores_val_type } from './writable_stores_val_type';
+import type { Writable } from './lib';
 /**
  * Sets each store with val
  */
-export declare function clear_store<I extends writable_stores_type<unknown>>(stores: I, val: writable_stores_val_type<I>): void;
+export declare function clear_store<I extends unknown>(stores: Writable<I> | Writable<I>[], val: I): void;
 export declare const clear__store: typeof clear_store;