+++ clear_store.ts
@@ -1,15 +1,13 @@
 import { _wrap_a1 } from '@ctx-core/function'
 import { each } from '@ctx-core/array'
-import type { writable_stores_type } from './writable_stores_type'
-import type { writable_stores_val_type } from './writable_stores_val_type'
+import type { Writable } from './lib'
 /**
  * Sets each store with val
  */
-export function clear_store<I extends writable_stores_type<unknown>>(
-	stores:I,
-	val: writable_stores_val_type<I>
+export function clear_store<I extends unknown>(
+	stores:Writable<I>|Writable<I>[], val:I
 ) {
-	const store_a1 = _wrap_a1<I>(stores)
-	each(store_a1, store=>store.set(val))
+	const store_a1 = _wrap_a1<Writable<I>>(stores)
+	each(store_a1, (store:Writable<I>)=>store.set(val))
 }
 export const clear__store = clear_store