+++ derived__spread.ts
@@ -1,8 +1,0 @@
-import { _spread } from '@ctx-core/array'
-import { derived } from './lib'
-import type { Stores } from './Stores'
-export function derived__spread<S extends Stores<unknown>, T extends unknown>(
-	stores: S, fn, initial_value?
-) {
-	return derived<S, T>(stores, _spread(fn), initial_value)
-}