+++ multi_subscribe.ts
@@ -1,21 +1,22 @@
 import { each, map } from '@ctx-core/array'
 import { get } from './get'
 import { subscribe } from './subscribe'
-import type { Readable } from './lib'
 import type { Unsubscriber } from './Unsubscriber'
+import type { Readable } from './lib'
 /**
  * Subscribes to multiple stores. The subscriber fn is called when any of the store_a1 changes.
  */
-export function subscribe__multi<I extends unknown>(
+export function multi_subscribe<I extends unknown>(
 	store_a1:Readable<I>[],
-	fn:((store_a1:Readable<I>[])=>void)
+	fn:((store_a1:I[])=>void)
 ) {
 	const unsubscribe_a1 =
 		map<Readable<I>, Unsubscriber>(store_a1,
-			(store, i)=>subscribe<I>(store,
+			(store, i)=>subscribe(
+				store,
 				$store=>invoke($store, i)
 			))
-	return ()=>each<Unsubscriber>(unsubscribe_a1, unsubscribe => unsubscribe())
+	return ()=>each<Unsubscriber>(unsubscribe_a1, unsubscribe=>unsubscribe())
 	function invoke($i_store, i) {
 		const all_$store_a1 =
 			map(store_a1,
@@ -27,3 +28,4 @@ export function subscribe__multi<I extends unknown>(
 		fn(all_$store_a1)
 	}
 }
+export const subscribe__multi = multi_subscribe