+++ readable_infer_type.ts
@@ -1,0 +1,3 @@
+import type { Readable } from './lib'
+export type readable_infer_type<I extends Readable<unknown>> =
+	I extends (infer O) ? O : never