+++ spread_derived.ts
@@ -1,0 +1,9 @@
+import { _spread } from '@ctx-core/array'
+import { derived } from './lib'
+import type { Stores } from './Stores'
+export function spread_derived<I extends Stores<unknown>, O extends unknown>(
+	stores: I, fn, initial_value?
+) {
+	return derived<I, O>(stores, _spread(fn), initial_value)
+}
+export const derived__spread = spread_derived